@@ -1,14 +1,16 @@
 import React from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
-import { Main, Login, OrderForm, ViewOrders} from '../components';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { Main, Login, OrderForm, ViewOrders } from '../components';
 
 const AppRouter = (props) => {
   return (
     <Router>
-      <Route path="/" exact component={Main} />
-      <Route path="/login" exact component={Login} />
-      <Route path="/order" exact component={OrderForm} />
-      <Route path="/view-orders" exact component={ViewOrders} />
+      <Routes>
+        <Route path="/" exact element={<Main/>} />
+        <Route path="/login" exact element={<Login/>} />
+        <Route path="/order" exact element={<OrderForm/>} />
+        <Route path="/view-orders" exact element={<ViewOrders/>} />
+      </Routes>
     </Router>
   );
 }
