@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import LoginForm from './login-form/loginForm';
+import withNavigateHook from './withNavigateHook';
 import './login.css';
 
 class Login extends Component {
@@ -9,11 +10,11 @@ class Login extends Component {
       <div className="main-body">
         <h1 className="text-center">Login Screen</h1>
         <div className="d-flex justify-content-center mt-5">
-          <LoginForm onLogin={() => {this.props.history.push('/view-orders')}}/>
+          <LoginForm onLogin={() => {this.props.navigation('/view-orders')}}/>
         </div>
       </div>
     )
   }
 }
 
-export default Login;
+export default withNavigateHook(Login);
