@@ -2,7 +2,7 @@ import { LOGIN, LOGOUT } from '../actions/types'
 
 const INITIAL_STATE = { email: null, token: null };
 
-export default (state = INITIAL_STATE, action) => {
+const authReducer = (state = INITIAL_STATE, action) => {
     switch (action.type) {
         case LOGIN:
             return { ...state, email: action.payload.login, token: action.payload.token }
@@ -12,3 +12,5 @@ export default (state = INITIAL_STATE, action) => {
             return state;
     }
 }
+
+export default authReducer;
